Implement a simple Contents.gz based icon finder
This icon finder code uses zgrep on Contents.gz files to find a matching
icon.
Ugly, but necessary since many packages ship their icons in a -data
package. Using zgrep turned out to be much faster than doing this in
Python, and it is also memory-efficient.
Still, I hope we can solve this issue in a smarter way in future.