Allow config to override the default icon-theme selection
Just use the `useIconTheme` to enforce a specific icon theme per
distribution suite.
This is useful for some 3rd-party repositories, which don't have the
default icon themes present, but ship with an own icon theme in the
repositories which can be used instead.
Please doN't never use this for full distribution repositories!

This resolves issue #6.