Honor "index.theme" manifest for XDG icon themes
This makes the generator read icon themes.
The current implementation is still horribly inefficient and needs some
more work to make it more useful and to clean up the code a bit, but it
is generally usable.

This should resolve issue #7, but needs a bit more testing to verify
that this change does not break support for existing themes. Also, the
code cleanup mentioned above should happen before this can go live at
Debian.